update CoC to version 2
The previous version of the CoC contained language that was too specific to contributions to code repositories. Moreover, there was no clear definition of what the consequences for those who violated the CoC were beyond the scope of code. This version of the CoC fills the gaps a lot better.